requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.18
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
